PIM-3832: rename converters
