+++ src/Pim/Bundle/BaseConnectorBundle/Processor/Denormalization/Converter/Flat/AttributeOptionToStandardConverter.php
@@ -1,22 +1,21 @@
 <?php
 
-namespace Pim\Bundle\BaseConnectorBundle\Reader\File\Converter;
+namespace Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\Converter\Flat;
 
+use Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\Converter\StandardArrayConverterInterface;
 use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 use Symfony\Component\Validator\Exception\InvalidOptionsException;
 
 /**
- * Attribute Option CSV Converter
+ * Attribute Option Flat Converter
  *
  * @author    Nicolas Dupont <nicola@akeneo.com>
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * TODO: rename to ArrayConverter + use Flat or Structured
  */
-class CsvAttributeOptionToStandardConverter implements StandardFormatConverterInterface
+class AttributeOptionToStandardConverter implements StandardArrayConverterInterface
 {
     /** @var LocaleRepositoryInterface */
     protected $localeRepository;
@@ -56,11 +55,11 @@ class CsvAttributeOptionToStandardConverter implements StandardFormatConverterIn
      *
      * @param array $item Representing a flat attribute option
      *
-     * @return array structured product
+     * @return array structured item
      *
      * @throws InvalidOptionsException
      */
-    public function convert($item)
+    public function convert(array $item)
     {
         $optionResolver = $this->createOptionsResolver();
         $resolvedItem = $optionResolver->resolve($item);
