+++ src/Pim/Bundle/BaseConnectorBundle/Processor/Denormalization/Converter/StandardArrayConverterInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\BaseConnectorBundle\Reader\File\Converter;
+namespace Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\Converter;
 
 /**
  * Standard converter interface, convert a format to the standard one
@@ -8,15 +8,13 @@ namespace Pim\Bundle\BaseConnectorBundle\Reader\File\Converter;
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * TODO: rename to ArrayConverter
  */
-interface StandardFormatConverterInterface
+interface StandardArrayConverterInterface
 {
     /**
-     * @param mixed $data
+     * @param array $data
      *
      * @return array
      */
-    public function convert($data);
+    public function convert(array $data);
 }
