+++ src/Pim/Bundle/BaseConnectorBundle/Resources/config/processors.yml
@@ -204,7 +204,6 @@ services:
         class: %pim_base_connector.denormalization.array_converter.structured.attribute_option.class%
 
     # Normalization
-
     pim_base_connector.processor.normalization.csv_group:
         class: %pim_base_connector.processor.csv_serializer.homogeneous.class%
         parent: pim_base_connector.processor.csv_serializer
